Renames library to index
